requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
google-genai>=0.8.0
streamlit>=1.35.0
streamlit-webrtc>=0.47.0
av>=11.0.0